Implement static canonicalize method

+++ lib/Synthesizer/estimator.h
@@ -21,7 +21,7 @@ public:
 private:
     Analyzer *m_analyzer;
     int m_nexception;
-    QVector<int> m_sums;
+    QVector<unsigned long long> m_sums;
     GiNaC::realsymbol m_s;
     GiNaC::realsymbol m_inf;
     QVector<bool> m_requirements;
